Separate and implement preset manager
